+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 - Fix issue where all morph information for an affix rule were being reported,
   rather than just that for the relevant pattern. Fixes
   [#73](https://github.com/pluots/zspell/issues/73).
-- Rewrite the `.dic` file parser
+- Rewrite the `.dic` file parser.
+- Improve handling of morphological information in dictionary files.
 
 ### Removed